refactor #60 : Extract chat room deletion logic into a new service
- Created ChatRoomDeletionService to handle the deletion of chat rooms independently.
- Updated ChatRoomBatchService to utilize the new service for deleting chat rooms, improving code organization and transaction management.
